+++ udemy/complete-nodejs-dev/task-manager/package.json
@@ -28,6 +28,7 @@
   "devDependencies": {
     "env-cmd": "^8.0.2",
     "jest": "^23.6.0",
-    "nodemon": "^1.18.9"
+    "nodemon": "^1.18.9",
+    "supertest": "^3.4.1"
   }
 }
